Pass in player instead of the player's sign into the game move.
This introduces coupling between the game and player objects,
but I decided that it was appropriate for the game to know about
its players, rather than just about the signs. I don't see any down
sides to this coupling within the scope of this game. The upside
is that the this allows the game to directly congratulate the winner
instead of us having to figure out the correct winner outside
the scope of the game.